refactor : Button -> Modal로 이름 변경
#220
