+++ src/pages/admin/LibraryManage/LibraryManage.tsx
@@ -7,8 +7,8 @@ import { ChildComponent } from '@components/Table/StandardTable.interface';
 import { columns, rows } from '@mocks/LibraryManageApi';
 
 import SearchSection from '@components/Section/SearchSection';
-import AddBookButton from './Button/AddBookButton';
-import DeleteBookButton from './Button/DeleteBookButton';
+import AddBookModal from './Modal/AddBookModal';
+import DeleteBookModal from './Modal/DeleteBookModal';
 
 const LibraryManage = () => {
   const childComponent = ({ key, value }: ChildComponent<BookListInfo>) => {
@@ -26,8 +26,8 @@ const LibraryManage = () => {
       <div className="mb-5 flex w-full items-center justify-between">
         <SearchSection />
         <div className="flex space-x-2">
-          <AddBookButton />
-          <DeleteBookButton />
+          <AddBookModal />
+          <DeleteBookModal />
           <OutlinedButton>대출 관리</OutlinedButton>
           <OutlinedButton>반납 관리</OutlinedButton>
         </div>
